Refs #124. Moved drawing graphic blip testing to its own module
